feat: move calculate* methods from WTS to SnapEngine (Phase B)
9 calculate methods moved: calculateSnapToAppRule, calculateSnapToLastZone,
calculateSnapToEmptyZone, calculateRestoreFromSession, calculateRotation,
calculateSnapAllWindowEntries, calculateResnapFromPreviousLayout,
calculateResnapFromCurrentAssignments, calculateResnapFromAutotileOrder.

New files: snapengine/calculate.cpp (auto-snap chain),
snapengine/resnap_calc.cpp (resnap/rotation/snap-all).

WTS gains public accessors: effectReportedWindows(), takeResnapBuffer(),
and promoted findEmptyZoneInLayout/sortZonesByNumber/buildZonePositionMap/
resolveEffectiveScreenId/resolveZoneGeometry to public.

All callers (SnapEngine, WTA snaprestore, WTA navigation, 4 test files)
updated to call through SnapEngine instead of WTS.

-1064/+128 lines — WTS shrinks significantly.
