+++ src/core/windowtrackingservice/navigation.cpp
@@ -185,114 +185,6 @@ QRect WindowTrackingService::multiZoneGeometry(const QStringList& zoneIds, const
     return combined.toAlignedRect();
 }
 
-QVector<ZoneAssignmentEntry> WindowTrackingService::calculateRotation(bool clockwise, const QString& screenFilter) const
-{
-    QVector<ZoneAssignmentEntry> result;
-
-    // Group snapped windows by screen so each screen rotates independently
-    // using its own per-screen layout (not the global active layout)
-    QHash<QString, QVector<QPair<QString, QString>>> windowsByScreen; // screenId -> [(windowId, primaryZoneId)]
-    for (auto it = m_windowZoneAssignments.constBegin(); it != m_windowZoneAssignments.constEnd(); ++it) {
-        // User-initiated snap commands override floating state.
-        // Floating windows are unsnapped (not in m_windowZoneAssignments), so this
-        // is a no-op, but we remove the check for consistency across all snap paths.
-
-        const QStringList& zoneIdList = it.value();
-        if (zoneIdList.isEmpty()) {
-            continue;
-        }
-
-        QString screenId = m_windowScreenAssignments.value(it.key());
-
-        // When a screen filter is set, only include windows on that screen
-        if (!screenFilter.isEmpty() && !Phosphor::Screens::ScreenIdentity::screensMatch(screenId, screenFilter)) {
-            continue;
-        }
-
-        windowsByScreen[screenId].append({it.key(), zoneIdList.first()});
-    }
-
-    // Process each screen independently
-    for (auto screenIt = windowsByScreen.constBegin(); screenIt != windowsByScreen.constEnd(); ++screenIt) {
-        const QString& screenId = screenIt.key();
-
-        // Get the layout assigned to THIS screen (not the global active layout)
-        PhosphorZones::Layout* layout = m_layoutManager->resolveLayoutForScreen(screenId);
-        if (!layout || layout->zoneCount() < 2) {
-            continue;
-        }
-
-        // Get zones sorted by zone number
-        QVector<PhosphorZones::Zone*> zones = layout->zones();
-        sortZonesByNumber(zones);
-
-        // Build zone ID -> index map (with and without braces for format-agnostic matching)
-        QHash<QString, int> zoneIdToIndex;
-        for (int i = 0; i < zones.size(); ++i) {
-            zoneIdToIndex[zones[i]->id().toString()] = i;
-        }
-
-        // Find zone indices for windows on this screen
-        QVector<QPair<QString, int>> windowZoneIndices;
-        for (const auto& windowEntry : screenIt.value()) {
-            const QString& windowId = windowEntry.first;
-            const QString& storedZoneId = windowEntry.second;
-            int zoneIndex = -1;
-
-            // Try direct match first
-            if (zoneIdToIndex.contains(storedZoneId)) {
-                zoneIndex = zoneIdToIndex.value(storedZoneId);
-            } else {
-                // Try matching by parsing as UUID (handles format differences)
-                QUuid storedUuid = QUuid::fromString(storedZoneId);
-                if (!storedUuid.isNull()) {
-                    for (int i = 0; i < zones.size(); ++i) {
-                        if (zones[i]->id() == storedUuid) {
-                            zoneIndex = i;
-                            break;
-                        }
-                    }
-                }
-            }
-
-            if (zoneIndex >= 0) {
-                windowZoneIndices.append({windowId, zoneIndex});
-            } else {
-                qCDebug(lcCore) << "Window" << windowId << "has zone ID" << storedZoneId
-                                << "not found in layout for screen" << screenId << "- skipping rotation";
-            }
-        }
-
-        // Resolve physical screen for zone geometry calculation
-        QScreen* screen = (m_screenManager ? m_screenManager->physicalQScreenFor(screenId)
-                                           : Utils::findScreenAtPosition(QPoint(0, 0)));
-        if (!screen) {
-            continue;
-        }
-
-        // Calculate rotated positions within this screen's zones
-        for (const auto& pair : windowZoneIndices) {
-            int currentIdx = pair.second;
-            int targetIdx =
-                clockwise ? (currentIdx + 1) % zones.size() : (currentIdx - 1 + zones.size()) % zones.size();
-
-            PhosphorZones::Zone* sourceZone = zones[currentIdx];
-            PhosphorZones::Zone* targetZone = zones[targetIdx];
-            QRect geo = GeometryUtils::getZoneGeometryForScreen(m_screenManager, targetZone, screen, screenId, layout,
-                                                                m_settings);
-
-            if (geo.isValid()) {
-                ZoneAssignmentEntry entry;
-                entry.windowId = pair.first;
-                entry.sourceZoneId = sourceZone->id().toString();
-                entry.targetZoneId = targetZone->id().toString();
-                entry.targetGeometry = geo;
-                result.append(entry);
-            }
-        }
-    }
-
-    return result;
-}
+// calculateRotation moved to SnapEngine (src/snap/snapengine/resnap_calc.cpp).
 
 } // namespace PlasmaZones
