+++ src/core/windowtrackingservice/resnap.cpp
@@ -21,118 +21,7 @@
 
 namespace PlasmaZones {
 
-QVector<ZoneAssignmentEntry> WindowTrackingService::calculateResnapFromPreviousLayout()
-{
-    QVector<ZoneAssignmentEntry> result;
-    if (m_resnapBuffer.isEmpty()) {
-        qCDebug(lcCore) << "calculateResnapFromPreviousLayout: buffer is empty";
-        return result;
-    }
-    qCDebug(lcCore) << "calculateResnapFromPreviousLayout: buffer has" << m_resnapBuffer.size() << "entries";
-
-    // Helper: create a "__restore__" entry that tells the KWin effect to move
-    // the window back to its pre-tile geometry instead of snapping it to a zone.
-    auto tryAppendRestore = [this, &result](const ResnapEntry* entry) {
-        auto preTile = preTileGeometry(entry->windowId);
-        if (!preTile) {
-            // Second attempt: look up by current class so a renamed window
-            // still finds its persisted pre-tile geometry.
-            preTile = preTileGeometry(currentAppIdFor(entry->windowId));
-        }
-        if (preTile && preTile->isValid()) {
-            ZoneAssignmentEntry restoreEntry;
-            restoreEntry.windowId = entry->windowId;
-            restoreEntry.sourceZoneId = QString();
-            restoreEntry.targetZoneId = QString(RestoreSentinel);
-            restoreEntry.targetGeometry = *preTile;
-            result.append(restoreEntry);
-        } else {
-            qCDebug(lcCore) << "No pre-tile geometry for excess window" << entry->windowId
-                            << "- window will remain at stale position";
-        }
-    };
-
-    // Group resnap entries by screen so each screen uses its own layout
-    QHash<QString, QVector<const ResnapEntry*>> entriesByScreen;
-    for (const ResnapEntry& entry : m_resnapBuffer) {
-        entriesByScreen[entry.screenId].append(&entry);
-    }
-
-    for (auto screenIt = entriesByScreen.constBegin(); screenIt != entriesByScreen.constEnd(); ++screenIt) {
-        const QString& screenId = screenIt.key();
-
-        // Get the layout assigned to this screen (not the global active layout)
-        PhosphorZones::Layout* newLayout = m_layoutManager->resolveLayoutForScreen(screenId);
-        if (!newLayout || newLayout->zoneCount() == 0) {
-            continue;
-        }
-
-        QVector<PhosphorZones::Zone*> newZones = newLayout->zones();
-        sortZonesByNumber(newZones);
-        const int newZoneCount = newZones.size();
-
-        for (const ResnapEntry* entry : screenIt.value()) {
-            // Multi-zone windows: map zone positions to the new layout.
-            // Only include positions that exist in the new layout — don't cycle
-            // excess positions (e.g. zone 3 shouldn't map to zone 1 when going
-            // from a 3-zone to a 2-zone layout).
-            if (entry->allZonePositions.size() > 1) {
-                QStringList targetZoneIds;
-                for (int pos : entry->allZonePositions) {
-                    if (pos > newZoneCount)
-                        continue; // skip positions beyond new layout
-                    PhosphorZones::Zone* z = newZones.value(pos - 1, nullptr);
-                    if (z)
-                        targetZoneIds.append(z->id().toString());
-                }
-                if (!targetZoneIds.isEmpty()) {
-                    QRect geo = resolveZoneGeometry(targetZoneIds, screenId);
-                    if (geo.isValid()) {
-                        ZoneAssignmentEntry assignEntry;
-                        assignEntry.windowId = entry->windowId;
-                        assignEntry.sourceZoneId = QString();
-                        assignEntry.targetZoneId = targetZoneIds.first();
-                        if (targetZoneIds.size() > 1)
-                            assignEntry.targetZoneIds = targetZoneIds;
-                        assignEntry.targetGeometry = geo;
-                        result.append(assignEntry);
-                    }
-                } else {
-                    // ALL zone positions exceed the new layout — restore to pre-tile geometry.
-                    tryAppendRestore(entry);
-                }
-            } else {
-                // Single-zone: map 1:1 by position (1->1, 2->2).
-                // Windows whose position exceeds the new zone count are restored to
-                // their pre-tile geometry — a window snapped to zone 3 should go back
-                // to its original size when switching to a 2-zone layout.
-                if (entry->zonePosition > newZoneCount) {
-                    tryAppendRestore(entry);
-                    continue;
-                }
-                PhosphorZones::Zone* targetZone = newZones.value(entry->zonePosition - 1, nullptr);
-                if (!targetZone) {
-                    continue;
-                }
-
-                QRect geo = zoneGeometry(targetZone->id().toString(), entry->screenId);
-                if (!geo.isValid()) {
-                    continue;
-                }
-
-                ZoneAssignmentEntry assignEntry;
-                assignEntry.windowId = entry->windowId;
-                assignEntry.sourceZoneId = QString();
-                assignEntry.targetZoneId = targetZone->id().toString();
-                assignEntry.targetGeometry = geo;
-                result.append(assignEntry);
-            }
-        }
-    }
-
-    m_resnapBuffer.clear();
-    return result;
-}
+// calculateResnapFromPreviousLayout moved to SnapEngine (src/snap/snapengine/resnap_calc.cpp).
 
 void WindowTrackingService::populateResnapBufferForAllScreens(const QSet<QString>& excludeScreens,
                                                               const QSet<QString>& includeScreens)
@@ -213,216 +102,8 @@ void WindowTrackingService::populateResnapBufferForAllScreens(const QSet<QString
     }
 }
 
-QVector<ZoneAssignmentEntry>
-WindowTrackingService::calculateResnapFromCurrentAssignments(const QString& screenFilter) const
-{
-    QVector<ZoneAssignmentEntry> result;
-
-    for (auto it = m_windowZoneAssignments.constBegin(); it != m_windowZoneAssignments.constEnd(); ++it) {
-        const QString& windowId = it.key();
-        const QStringList& zoneIds = it.value();
-        if (zoneIds.isEmpty()) {
-            continue;
-        }
-        // Skip ALL floating windows. Floating state persists across mode
-        // toggles — if the user floated a window (in either mode), it stays
-        // floating and should not be resnapped to a zone.
-        if (isWindowFloating(windowId)) {
-            continue;
-        }
-
-        QString screenId = m_windowScreenAssignments.value(windowId);
-        if (!screenFilter.isEmpty()) {
-            // If the filter is a virtual screen ID, require exact equality —
-            // screensMatch already enforces that distinct VS IDs (and VS vs
-            // physical) never match. If the filter is a physical screen ID,
-            // match windows stored on that physical screen OR on any of its
-            // virtual children — belongsToPhysicalScreen handles both cases.
-            const bool match = PhosphorIdentity::VirtualScreenId::isVirtual(screenFilter)
-                ? Phosphor::Screens::ScreenIdentity::screensMatch(screenId, screenFilter)
-                : Utils::belongsToPhysicalScreen(screenId, screenFilter);
-            if (!match) {
-                continue;
-            }
-        }
-
-        // NOTE: no in-loop mode filter here. Callers are responsible for
-        // passing a screenFilter that targets only snap-mode screens, and
-        // for iterating per-screen when multiple screens need resnapping.
-        // The dispatch layer (ScreenModeRouter via WindowTrackingAdaptor)
-        // enforces that contract at the entry point so the engine and its
-        // service helpers can stay pure and free of defensive checks.
-
-        QRect geo = resolveZoneGeometry(zoneIds, screenId);
-        if (!geo.isValid()) {
-            continue;
-        }
-
-        ZoneAssignmentEntry entry;
-        entry.windowId = windowId;
-        entry.sourceZoneId = QString();
-        entry.targetZoneId = zoneIds.first();
-        if (zoneIds.size() > 1)
-            entry.targetZoneIds = zoneIds;
-        entry.targetGeometry = geo;
-        // Stamp the authoritative target screen so processBatchEntries skips
-        // re-derivation from geometry.center(). If the layout's geometry ever
-        // resolved to a stale fallback (e.g. a transient cache miss), the
-        // re-derivation could land in a sibling VS and clobber the stored
-        // assignment. Trust the source screen we already read above.
-        entry.targetScreenId = screenId;
-        result.append(entry);
-    }
-
-    qCInfo(lcCore) << "Resnap from current assignments:" << result.size() << "windows"
-                   << "(total zone assignments:" << m_windowZoneAssignments.size() << ")"
-                   << (screenFilter.isEmpty() ? QStringLiteral("(all screens)")
-                                              : QStringLiteral("(screen: %1)").arg(screenFilter));
-    if (result.isEmpty() && !m_windowZoneAssignments.isEmpty() && lcCore().isDebugEnabled()) {
-        auto screenMatches = [&](const QString& screen) {
-            if (screenFilter.isEmpty())
-                return true;
-            return PhosphorIdentity::VirtualScreenId::isVirtual(screenFilter)
-                ? Phosphor::Screens::ScreenIdentity::screensMatch(screen, screenFilter)
-                : Utils::belongsToPhysicalScreen(screen, screenFilter);
-        };
-        for (auto it = m_windowZoneAssignments.constBegin(); it != m_windowZoneAssignments.constEnd(); ++it) {
-            QString screen = m_windowScreenAssignments.value(it.key());
-            bool floating = isWindowFloating(it.key());
-            QRect geo = it.value().isEmpty() ? QRect() : resolveZoneGeometry(it.value(), screen);
-            qCDebug(lcCore) << "  skipped:" << it.key() << "zones=" << it.value() << "screen=" << screen
-                            << "floating=" << floating << "geoValid=" << geo.isValid()
-                            << "screenMatch=" << screenMatches(screen);
-        }
-    }
-    return result;
-}
-
-QVector<ZoneAssignmentEntry>
-WindowTrackingService::calculateResnapFromAutotileOrder(const QStringList& autotileWindowOrder,
-                                                        const QString& screenId) const
-{
-    QVector<ZoneAssignmentEntry> result;
-
-    if (autotileWindowOrder.isEmpty()) {
-        return result;
-    }
-
-    PhosphorZones::Layout* layout = m_layoutManager->resolveLayoutForScreen(screenId);
-    if (!layout || layout->zoneCount() == 0) {
-        qCWarning(lcCore) << "calculateResnapFromAutotileOrder: no layout for screen" << screenId;
-        return result;
-    }
-
-    QVector<PhosphorZones::Zone*> zones = layout->zones();
-    sortZonesByNumber(zones);
-
-    const int zoneCount = zones.size();
-    const int windowCount = autotileWindowOrder.size();
-
-    // Cap at zoneCount: excess windows beyond available zones would stack on top of each
-    // other with cycling. Instead, leave them unassigned (they stay where they are).
-    if (windowCount > zoneCount) {
-        qCWarning(lcCore) << "calculateResnapFromAutotileOrder:" << windowCount << "windows but only" << zoneCount
-                          << "zones on screen" << screenId << "- excess" << (windowCount - zoneCount)
-                          << "windows will not be resnapped";
-    }
-
-    // Build a lookup from zone ID → zone pointer for original-zone restoration
-    QHash<QString, PhosphorZones::Zone*> zoneById;
-    for (PhosphorZones::Zone* z : zones) {
-        zoneById[z->id().toString()] = z;
-    }
-
-    // Track which zones have been claimed by original-assignment restoration
-    // so positional fallback doesn't double-assign.
-    QSet<int> claimedZoneIndices;
-    // Track windows placed in the first pass for O(1) lookup in the second pass
-    QSet<QString> placedWindowIds;
-
-    // First pass: restore windows to their ORIGINAL zone assignment (pre-autotile).
-    // m_windowZoneAssignments preserves zone IDs from before autotile was activated.
-    // This ensures a window in zone 3 returns to zone 3, not whatever autotile
-    // position it ended up in.
-    for (int i = 0; i < std::min(windowCount, zoneCount); ++i) {
-        const QString& windowId = autotileWindowOrder.at(i);
-        const QStringList& savedZones = m_windowZoneAssignments.value(windowId);
-
-        if (savedZones.isEmpty())
-            continue;
-
-        // Restore to ALL saved zone assignments (supports multi-zone spans).
-        // Verify all zones still exist in the restored layout.
-        QStringList validZoneIds;
-        QList<int> validZoneIndices;
-        for (const QString& zid : savedZones) {
-            PhosphorZones::Zone* z = zoneById.value(zid);
-            if (!z)
-                continue;
-            int idx = zones.indexOf(z);
-            if (idx < 0 || claimedZoneIndices.contains(idx))
-                continue;
-            validZoneIds.append(zid);
-            validZoneIndices.append(idx);
-        }
-        if (validZoneIds.isEmpty())
-            continue;
-
-        // Compute geometry: combined for multi-zone, single for normal
-        QRect geo = resolveZoneGeometry(validZoneIds, screenId);
-
-        if (geo.isValid()) {
-            ZoneAssignmentEntry entry;
-            entry.windowId = windowId;
-            entry.sourceZoneId = QString();
-            entry.targetZoneId = validZoneIds.first();
-            if (validZoneIds.size() > 1)
-                entry.targetZoneIds = validZoneIds;
-            entry.targetGeometry = geo;
-            result.append(entry);
-            placedWindowIds.insert(windowId);
-            for (int idx : validZoneIndices)
-                claimedZoneIndices.insert(idx);
-        }
-    }
-
-    // Second pass: windows without a valid original zone get positional fallback
-    // (autotile order position → first unclaimed zone).
-    for (int i = 0; i < std::min(windowCount, zoneCount); ++i) {
-        const QString& windowId = autotileWindowOrder.at(i);
-
-        // Skip if already placed by original-zone restoration
-        if (placedWindowIds.contains(windowId))
-            continue;
-
-        // Find next unclaimed zone
-        int zoneIdx = -1;
-        for (int z = 0; z < zoneCount; ++z) {
-            if (!claimedZoneIndices.contains(z)) {
-                zoneIdx = z;
-                break;
-            }
-        }
-        if (zoneIdx < 0)
-            break; // No more zones available
-
-        PhosphorZones::Zone* targetZone = zones.at(zoneIdx);
-        QRect geo = zoneGeometry(targetZone->id().toString(), screenId);
-
-        if (geo.isValid()) {
-            ZoneAssignmentEntry entry;
-            entry.windowId = windowId;
-            entry.sourceZoneId = QString();
-            entry.targetZoneId = targetZone->id().toString();
-            entry.targetGeometry = geo;
-            result.append(entry);
-            claimedZoneIndices.insert(zoneIdx);
-        }
-    }
-
-    qCInfo(lcCore) << "Resnap from autotile order:" << result.size() << "windows for screen" << screenId;
-    return result;
-}
+// calculateResnapFromCurrentAssignments, calculateResnapFromAutotileOrder
+// moved to SnapEngine (src/snap/snapengine/resnap_calc.cpp).
 
 QStringList WindowTrackingService::buildZoneOrderedWindowList(const QString& screenId) const
 {
@@ -490,67 +171,7 @@ QStringList WindowTrackingService::buildZoneOrderedWindowList(const QString& scr
     return result;
 }
 
-QVector<ZoneAssignmentEntry> WindowTrackingService::calculateSnapAllWindows(const QStringList& windowIds,
-                                                                            const QString& screenId) const
-{
-    QVector<ZoneAssignmentEntry> result;
-
-    PhosphorZones::Layout* layout = m_layoutManager->resolveLayoutForScreen(screenId);
-    if (!layout || layout->zoneCount() == 0) {
-        return result;
-    }
-
-    QVector<PhosphorZones::Zone*> zones = layout->zones();
-    sortZonesByNumber(zones);
-
-    // Filter occupancy by the current virtual desktop so windows parked on other
-    // desktops don't make zones appear occupied on the current-desktop batch snap.
-    const int desktopFilter = m_virtualDesktopManager ? m_virtualDesktopManager->currentDesktop() : 0;
-    QSet<QUuid> occupiedZoneIds = buildOccupiedZoneSet(screenId, desktopFilter);
-
-    // Resolve physical screen for zone geometry calculation
-    QScreen* screen =
-        (m_screenManager ? m_screenManager->physicalQScreenFor(screenId) : Utils::findScreenAtPosition(QPoint(0, 0)));
-    if (!screen) {
-        return result;
-    }
-
-    // Track zones we're assigning in this batch (to avoid double-assigning)
-    QSet<QUuid> batchOccupied = occupiedZoneIds;
-
-    for (const QString& windowId : windowIds) {
-        // Find the first unoccupied zone
-        PhosphorZones::Zone* targetZone = nullptr;
-        for (PhosphorZones::Zone* zone : zones) {
-            if (!batchOccupied.contains(zone->id())) {
-                targetZone = zone;
-                break;
-            }
-        }
-
-        if (!targetZone) {
-            // No more empty zones available
-            break;
-        }
-
-        QRect geo =
-            GeometryUtils::getZoneGeometryForScreen(m_screenManager, targetZone, screen, screenId, layout, m_settings);
-
-        if (geo.isValid()) {
-            ZoneAssignmentEntry entry;
-            entry.windowId = windowId;
-            entry.sourceZoneId = QString(); // Not previously snapped
-            entry.targetZoneId = targetZone->id().toString();
-            entry.targetGeometry = geo;
-            result.append(entry);
-
-            // Mark zone as occupied for subsequent iterations
-            batchOccupied.insert(targetZone->id());
-        }
-    }
-
-    return result;
-}
+// calculateSnapAllWindows moved to SnapEngine (src/snap/snapengine/resnap_calc.cpp).
 
 // ═══════════════════════════════════════════════════════════════════════════════
 // Resolution Change Handling
